backend/requirements.txt
Django>=4.2
djangorestframework
django-cors-headers
mysqlclient
djangorestframework-simplejwt
django-environ
python-dotenv
pillow
